Add files via upload
Understanding how the board for the game works. I created a function for the board game, and at each index, I inputted a character. The first index[0] is omitted so we skip that location and start at one from left to right and count up. We can put a character at whatever index "X' or "O".